Move commands into own subdir
