+++ src/commands/__tests__/fixtures/sanity-template-advanced/sanity-template.json
