+++ src/commands/__tests__/fixtures/sanity-template-advanced/template/package.json
