+++ src/commands/__tests__/fixtures/sanity-template-advanced/template/studio/package.json
