+++ src/commands/__tests__/fixtures/sanity-template-advanced/template/web/package.json
