+++ src/commands/__tests__/fixtures/sanity-template-empty/.gitignore
