+++ src/commands/__tests__/fixtures/sanity-template-minimal/sanity-template.json
