+++ src/commands/__tests__/fixtures/sanity-template-simple/sanity-template.json
