+++ src/commands/__tests__/fixtures/sanity-template-simple/template/package.json
