+++ src/commands/check.js
