+++ src/commands/utils/buildFile.js
