+++ src/commands/utils/replaceVars.js
