+++ src/commands/utils/watchFiles.js
